# Conflicts:
#	ams/ams-server/src/main/java/com/netease/arctic/ams/server/mapper/OptimizerMapper.java
#	ams/ams-server/src/main/java/com/netease/arctic/ams/server/service/impl/OptimizerService.java

+++ spark/v2.3/spark/src/test/java/com/netease/arctic/spark/SparkTestContext.java
@@ -1,4 +1,4 @@
-package com.netease.arctic.spark;/*
+/*
  * Licensed to the Apache Software Foundation (ASF) under one
  * or more contributor license agreements.  See the NOTICE file
  * distributed with this work for additional information
@@ -16,8 +16,9 @@ package com.netease.arctic.spark;/*
  * limitations under the License.
  */
 
+package com.netease.arctic.spark;
+
 import com.google.common.collect.Sets;
-import com.netease.arctic.CatalogMetaTestUtil;
 import com.netease.arctic.ams.api.CatalogMeta;
 import com.netease.arctic.ams.api.MockArcticMetastoreServer;
 import com.netease.arctic.ams.api.NoSuchObjectException;
@@ -26,13 +27,21 @@ import com.netease.arctic.ams.api.client.AmsClientPools;
 import com.netease.arctic.catalog.ArcticCatalog;
 import com.netease.arctic.catalog.CatalogLoader;
 import com.netease.arctic.data.ChangeAction;
+import com.netease.arctic.data.DataTreeNode;
+import com.netease.arctic.hive.HMSMockServer;
+import com.netease.arctic.hive.io.writer.AdaptHiveGenericTaskWriterBuilder;
 import com.netease.arctic.io.writer.GenericTaskWriters;
-import com.netease.arctic.spark.hive.SparkCatalogMetaTestUtil;
+import com.netease.arctic.io.writer.SortedPosDeleteWriter;
+import com.netease.arctic.op.OverwriteBaseFiles;
 import com.netease.arctic.table.ArcticTable;
 import com.netease.arctic.table.KeyedTable;
+import com.netease.arctic.table.LocationKind;
 import com.netease.arctic.table.TableIdentifier;
 import com.netease.arctic.table.UnkeyedTable;
+import com.netease.arctic.utils.FileUtil;
 import org.apache.commons.io.FileUtils;
+import org.apache.commons.lang.StringUtils;
+import org.apache.hadoop.hive.conf.HiveConf;
 import org.apache.iceberg.*;
 import org.apache.iceberg.data.GenericRecord;
 import org.apache.iceberg.data.Record;
@@ -44,7 +53,6 @@ import org.apache.iceberg.relocated.com.google.common.collect.Maps;
 import org.apache.iceberg.util.StructLikeMap;
 import org.apache.spark.SparkConf;
 import org.apache.spark.sql.Dataset;
-import org.apache.spark.sql.Encoder;
 import org.apache.spark.sql.Row;
 import org.apache.spark.sql.SparkSession;
 import org.apache.spark.sql.internal.SQLConf;
@@ -64,6 +72,7 @@ import java.time.OffsetDateTime;
 import java.time.ZoneOffset;
 import java.util.*;
 import java.util.concurrent.ConcurrentHashMap;
+import java.util.concurrent.atomic.AtomicInteger;
 import java.util.stream.Collectors;
 import java.util.stream.IntStream;
 
@@ -75,7 +84,6 @@ public class SparkTestContext extends ExternalResource {
   final static ConcurrentHashMap<String, ArcticCatalog> catalogs = new ConcurrentHashMap<>();
   private static final Logger LOG = LoggerFactory.getLogger(SparkTestBase.class);
 
-  protected static Map<String, String> additionSparkConfigs = Maps.newHashMap();
   protected static File testBaseDir = new File("unit_test_base_tmp");
   protected static File testSparkDir = new File(testBaseDir, "spark-warehouse");
   protected static File testArcticDir = new File(testBaseDir, "arctic");
@@ -84,13 +92,20 @@ public class SparkTestContext extends ExternalResource {
   protected static String amsUrl;
   protected static String catalogName;
   protected List<Object[]> rows;
+  static final File hmsDir = new File(testBaseDir, "hive");
+  protected static final HMSMockServer hms = new HMSMockServer(hmsDir);
 
-  public static
-  void cleanUpAdditionSparkConfigs() {
-    additionSparkConfigs.clear();
-  }
+  protected static final AtomicInteger arcticSetupCount = new AtomicInteger(0);
+  protected static final AtomicInteger hiveSetupCount = new AtomicInteger(0);
+  protected static final AtomicInteger sparkSetupCount = new AtomicInteger(0);
 
-  public static void setUpTestDirAndArctic() throws IOException {
+  public static Map<String, String> setUpTestDirAndArctic() throws IOException {
+    Map<String, String> configs = Maps.newHashMap();
+
+    if (arcticSetupCount.incrementAndGet() > 1) {
+      return configs;
+    }
+
     System.out.println("======================== start AMS  ========================= ");
     FileUtils.deleteQuietly(testBaseDir);
     testBaseDir.mkdirs();
@@ -100,14 +115,46 @@ public class SparkTestContext extends ExternalResource {
       ams.start();
     }
     amsUrl = "thrift://127.0.0.1:" + ams.port();
-
-    CatalogMeta arctic = SparkCatalogMetaTestUtil.createArcticCatalog(testArcticDir);
-    catalogName = arctic.getCatalogName();
-    ams.handler().createCatalog(arctic);
+    return configs;
   }
 
+  public static Map<String, String> setUpHMS() throws IOException, ClassNotFoundException {
+    Map<String, String> configs = Maps.newHashMap();
+    String hiveVersion = SparkTestContext.class.getClassLoader()
+        .loadClass("org.apache.hadoop.hive.metastore.HiveMetaStoreClient")
+        .getPackage()
+        .getImplementationVersion();
 
-  public static void setUpSparkSession() {
+
+    configs.put("spark.sql.catalogImplementation", "hive");
+    configs.put("hive.metastore.uris", "thrift://127.0.0.1:" + hms.getMetastorePort());
+    configs.put("spark.arctic.sql.delegate.enable", "true");
+    //hive.metastore.client.capability.check
+    configs.put("hive.metastore.client.capability.check", "false");
+
+    if (hiveSetupCount.incrementAndGet() > 1) {
+      return configs;
+    }
+
+    System.out.println("======================== start hive metastore ========================= ");
+    hms.hiveConf().set("datanucleus.autoCreateTables", "true");
+    hms.start();
+    HiveConf entries = hms.hiveConf();
+    CatalogMeta arctic_hive = ArcticCatalogMetaTestUtil.createArcticCatalog(testArcticDir, entries);
+    catalogName = arctic_hive.getCatalogName();
+    ams.handler().createCatalog(arctic_hive);
+
+    configs.put("spark.sql.arctic.catalog." + catalogName, ArcticCatalog.class.getName());
+    configs.put("spark.sql.arctic.catalog.type", "hive");
+    configs.put("spark.sql.arctic.catalog.url" , amsUrl + "/" + catalogName);
+    return configs;
+  }
+
+  public static void setUpSparkSession(Map<String, String> configs) {
+    if (sparkSetupCount.incrementAndGet() > 1) {
+      return;
+    }
+
     System.out.println("======================== set up spark session  ========================= ");
     Map<String, String> sparkConfigs = Maps.newHashMap();
 
@@ -117,14 +164,12 @@ public class SparkTestContext extends ExternalResource {
     sparkConfigs.put("spark.sql.warehouse.dir", testSparkDir.getAbsolutePath());
     sparkConfigs.put("spark.sql.extensions", ArcticSparkExtensions.class.getName());
     sparkConfigs.put("spark.testing.memory", "471859200");
+    sparkConfigs.put("spark.cores.max", "6");
+    sparkConfigs.put("spark.executor.cores", "2");
+    sparkConfigs.put("spark.default.parallelism", "12");
 
-    sparkConfigs.put("arctic.catalog." + catalogName, ArcticCatalog.class.getName());
-    sparkConfigs.put("arctic.catalog.type", "hive");
-    sparkConfigs.put("arctic.catalog.url", amsUrl + "/" + catalogName);
-
-    sparkConfigs.putAll(additionSparkConfigs);
+    sparkConfigs.putAll(configs);
     sparkConfigs.forEach(((k, v) -> System.out.println("--" + k + "=" + v)));
-
 
     SparkConf sparkconf = new SparkConf()
         .setAppName("test")
@@ -135,18 +180,31 @@ public class SparkTestContext extends ExternalResource {
     spark = SparkSession
         .builder()
         .config(sparkconf)
-        .enableHiveSupport()
         .getOrCreate();
     spark.sparkContext().setLogLevel("WARN");
   }
 
   public static void cleanUpAms() {
+    if (arcticSetupCount.decrementAndGet() > 0) {
+      return;
+    }
     System.out.println("======================== clean up AMS  ========================= ");
     ams.handler().cleanUp();
     AmsClientPools.cleanAll();
   }
 
-  public static void cleanUpSparkSession(){
+  public static void cleanUpHive() {
+    if (hiveSetupCount.decrementAndGet() > 0) {
+      return;
+    }
+    System.out.println("======================== stop hive metastore ========================= ");
+    hms.stop();
+  }
+
+  public static void cleanUpSparkSession() {
+    if (sparkSetupCount.decrementAndGet() > 0) {
+      return;
+    }
     System.out.println("======================== clean up spark session  ========================= ");
     spark.stop();
     spark.close();
@@ -175,9 +233,10 @@ public class SparkTestContext extends ExternalResource {
     }
   }
 
-  public static void writeBase(TableIdentifier identifier, List<Object[]> records) {
+  public static List<DataFile> writeBase(TableIdentifier identifier, List<Object[]> records) {
     KeyedTable table = SparkTestContext.catalog(identifier.getCatalog()).loadTable(identifier).asKeyedTable();
     long txId = table.beginTransaction(System.currentTimeMillis() + "");
+    List<DataFile> baseDataFiles = new ArrayList<>();
     try (TaskWriter<Record> writer = GenericTaskWriters.builderFor(table)
         .withTransactionId(txId)
         .buildBaseWriter()) {
@@ -189,10 +248,11 @@ public class SparkTestContext extends ExternalResource {
           throw new UncheckedIOException(e);
         }
       });
+      baseDataFiles.addAll(Arrays.asList(writer.complete().dataFiles()));
       AppendFiles appendFiles = table.baseTable().newAppend();
-      Arrays.stream(writer.complete().dataFiles())
-          .forEach(appendFiles::appendFile);
+      baseDataFiles.forEach(appendFiles::appendFile);
       appendFiles.commit();
+      return baseDataFiles;
     } catch (IOException e) {
       throw new RuntimeException(e);
     }
@@ -288,23 +348,13 @@ public class SparkTestContext extends ExternalResource {
       this.rows = new ArrayList<>();
       return ImmutableList.of();
     }
-    result.show();
+    result.show(false);
     this.rows = rows.stream()
         .map(row -> IntStream.range(0, row.size())
             .mapToObj(pos -> row.isNullAt(pos) ? null : row.get(pos))
             .toArray(Object[]::new)
         ).collect(Collectors.toList());
     return this.rows;
-  }
-
-  protected <T> List<T> sql(String query, Encoder<T> encoder, Object... args) throws RuntimeException {
-    MessageFormat format = new MessageFormat(query);
-    String sql = format.format(args);
-    if (args.length == 0) {
-      sql = query;
-    }
-    LOG.info("execute sql: " + sql);
-    return spark.sql(sql).as(encoder).collectAsList();
   }
 
   protected void assertEquals(String context, List<Object[]> expectedRows, List<Object[]> actualRows) {
@@ -338,5 +388,130 @@ public class SparkTestContext extends ExternalResource {
 
   protected static Timestamp quickTs(int day) {
     return Timestamp.valueOf(quickDateWithZone(day).toLocalDateTime());
+  }
+
+  public void assertDescResult(List<Object[]> rows, List<String> primaryKeys) {
+    boolean primaryKeysBlock = false;
+    List<String> descPrimaryKeys = Lists.newArrayList();
+    for (Object[] row : rows) {
+      if (StringUtils.equalsIgnoreCase("# Primary keys", row[0].toString())) {
+        primaryKeysBlock = true;
+      } else if (StringUtils.startsWith(row[0].toString(), "# ") && primaryKeysBlock) {
+        primaryKeysBlock = false;
+      } else if (primaryKeysBlock) {
+        descPrimaryKeys.add(row[0].toString());
+      }
+    }
+
+    Assert.assertEquals(primaryKeys.size(), descPrimaryKeys.size());
+    Assert.assertArrayEquals(
+        primaryKeys.stream().sorted().distinct().toArray(),
+        descPrimaryKeys.stream().sorted().distinct().toArray());
+  }
+
+  public void assertPartitionResult(List<Object[]> rows, List<String> partitionKey) {
+    boolean primaryKeysBlock = false;
+    List<String> descPrimaryKeys = Lists.newArrayList();
+    for (Object[] row : rows) {
+      if (StringUtils.equalsIgnoreCase("# Partitioning", row[0].toString())) {
+        primaryKeysBlock = true;
+      } else if (StringUtils.startsWith(row[0].toString(), "Part ") && primaryKeysBlock) {
+        descPrimaryKeys.add(row[1].toString());
+      }
+    }
+
+    Assert.assertEquals(partitionKey.size(), descPrimaryKeys.size());
+    Assert.assertArrayEquals(
+        partitionKey.stream().sorted().distinct().toArray(),
+        descPrimaryKeys.stream().sorted().distinct().toArray());
+  }
+
+  public void assertHiveDesc(List<Object[]> rows, List<String> cols, List<String> partitionKey) {
+    boolean colsBlock = true;
+    boolean partitionBlock = false;
+    List<String> descCols = Lists.newArrayList();
+    List<String> descPartitionKey = Lists.newArrayList();
+    for (Object[] row : rows) {
+      if (StringUtils.startsWith(row[0].toString(), "# Partition Infor")) {
+        partitionBlock = true;
+        colsBlock = false;
+      } else if (colsBlock) {
+        descCols.add(row[0].toString());
+      } else if (partitionBlock && !StringUtils.startsWith(row[0].toString(), "#")) {
+        descPartitionKey.add(row[0].toString());
+      }
+    }
+    Assert.assertArrayEquals(
+        cols.stream().sorted().distinct().toArray(),
+        descCols.stream().sorted().distinct().toArray());
+    Assert.assertArrayEquals(
+        partitionKey.stream().sorted().distinct().toArray(),
+        descPartitionKey.stream().sorted().distinct().toArray());
+  }
+
+  public List<DataFile> writeHive(ArcticTable table, LocationKind locationKind, List<Record> records)
+      throws IOException {
+    AdaptHiveGenericTaskWriterBuilder builder = AdaptHiveGenericTaskWriterBuilder
+        .builderFor(table)
+        .withTransactionId(table.isKeyedTable() ? 1L : null);
+
+    TaskWriter<Record> changeWrite = builder.buildWriter(locationKind);
+    for (Record record : records) {
+      changeWrite.write(record);
+    }
+    DataFile[] dataFiles = changeWrite.complete().dataFiles();
+    if (table.isKeyedTable()) {
+      KeyedTable keyedTable = table.asKeyedTable();
+      OverwriteBaseFiles overwriteBaseFiles = keyedTable.newOverwriteBaseFiles();
+      Arrays.stream(dataFiles).forEach(overwriteBaseFiles::addFile);
+      overwriteBaseFiles.withTransactionId(keyedTable.beginTransaction(System.currentTimeMillis() + ""));
+      overwriteBaseFiles.commit();
+    } else if (table.isUnkeyedTable()) {
+      UnkeyedTable unkeyedTable = table.asUnkeyedTable();
+      OverwriteFiles overwriteFiles = unkeyedTable.newOverwrite();
+      Arrays.stream(dataFiles).forEach(overwriteFiles::addFile);
+      overwriteFiles.commit();
+    } else {
+      throw new IllegalStateException("Table is neither keyed nor unkeyed");
+    }
+    return Arrays.asList(dataFiles);
+  }
+
+  public void adaptHiveInsertPosDeleteFiles(Long transactionId, List<DataFile> dataFiles, ArcticTable table)
+      throws IOException {
+    Map<StructLike, List<DataFile>> dataFilesPartitionMap =
+        new HashMap<>(dataFiles.stream().collect(Collectors.groupingBy(ContentFile::partition)));
+    List<DeleteFile> deleteFiles = new ArrayList<>();
+    for (Map.Entry<StructLike, List<DataFile>> dataFilePartitionMap : dataFilesPartitionMap.entrySet()) {
+      StructLike partition = dataFilePartitionMap.getKey();
+      List<DataFile> partitionFiles = dataFilePartitionMap.getValue();
+      Map<DataTreeNode, List<DataFile>> nodeFilesPartitionMap = new HashMap<>(partitionFiles.stream()
+          .collect(Collectors.groupingBy(dataFile ->
+              FileUtil.parseFileNodeFromFileName(dataFile.path().toString()))));
+      for (Map.Entry<DataTreeNode, List<DataFile>> nodeFilePartitionMap : nodeFilesPartitionMap.entrySet()) {
+        DataTreeNode key = nodeFilePartitionMap.getKey();
+        List<DataFile> nodeFiles = nodeFilePartitionMap.getValue();
+
+        // write pos delete
+        SortedPosDeleteWriter<Record> writer = AdaptHiveGenericTaskWriterBuilder
+            .builderFor(table)
+            .withTransactionId(transactionId).buildBasePosDeleteWriter(key.getMask(), key.getIndex(), partition);
+        for (DataFile nodeFile : nodeFiles) {
+          writer.delete(nodeFile.path(), 0);
+        }
+        deleteFiles.addAll(writer.complete());
+      }
+    }
+    if (table.isKeyedTable()) {
+      RowDelta rowDelta = table.asKeyedTable().baseTable().newRowDelta();
+      deleteFiles.forEach(rowDelta::addDeletes);
+      rowDelta.commit();
+    } else if (table.isUnkeyedTable()) {
+      RowDelta rowDelta = table.asUnkeyedTable().newRowDelta();
+      deleteFiles.forEach(rowDelta::addDeletes);
+      rowDelta.commit();
+    } else {
+      throw new IllegalStateException("Table is neither keyed nor unkeyed");
+    }
   }
 }
